# Fix injectRefreshLoader performance issue