--- a/lib/utils/injectRefreshLoader.js
+++ b/lib/utils/injectRefreshLoader.js
@@ -13,8 +13,8 @@ const path = require('path');
  */
 
 const resolvedLoader = require.resolve('../../loader');
-const reactRefreshPath = path.dirname(require.resolve('react-refresh'))
-const refreshUtilsPath = path.join(__dirname, '../runtime/RefreshUtils')
+const reactRefreshPath = path.dirname(require.resolve('react-refresh'));
+const refreshUtilsPath = path.join(__dirname, '../runtime/RefreshUtils');
 
 /**
  * Injects refresh loader to all JavaScript-like and user-specified files.