@JSerFeng 2022-09-27
Any idea on why `require.resolve` is not cached 🤔